Move the js-accordion init to global
