+++ fec/fec/static/js/global.js
@@ -1,9 +1,10 @@
 /**
  * Scripts used on every page of the site e.g. main nav, search, feedback
  */
+import { default as A11yDialog } from 'a11y-dialog';
+import { Accordion } from 'aria-accordion';
 import Glossary from 'glossary-panel/src/glossary.js';
 
-import { default as A11yDialog } from 'a11y-dialog';
 import { default as terms } from './data/terms.json' assert { type: 'json' };
 import Feedback from './modules/feedback.js';
 import SiteNav from './modules/site-nav.js';
@@ -12,6 +13,30 @@ import TOC from './modules/toc.js';
 import Typeahead from './modules/typeahead.js';
 
 $(function() {
+  // Initialize new accordions, including the .gov ribbon at the top of every page
+  $('.js-accordion').each(function() {
+    const contentPrefix = $(this).data('content-prefix') || 'accordion';
+    const openFirst = $(this).data('open-first');
+    const reflectStatic = $(this).data('reflect-static');
+    const selectors = {
+      trigger: '.js-accordion-trigger'
+    };
+    const opts = {
+      contentPrefix: contentPrefix,
+      openFirst: openFirst,
+      reflectStatic: reflectStatic,
+      collapseOthers:
+        window.location.href.indexOf('president/presidential-map') > 0
+          ? true
+          : false,
+      customHiding:
+        window.location.href.indexOf('president/presidential-map') > 0
+          ? true
+          : false
+    };
+    new Accordion(this, selectors, opts);
+  });
+
   /**
    * Glossary and "Table of Contents"
    */
